backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
transformers
torch
pillow
numpy
